requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
boto3==1.40.66
botocore==1.40.66
click==8.3.0
decorator==5.2.1
exceptiongroup==1.3.0
fastapi==0.121.0
h11==0.16.0
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
jmespath==1.0.1
moviepy==2.2.1
numpy==2.2.6
pillow==11.3.0
proglog==0.1.12
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
